매직 미사일 프로토 타입
파이어볼에 사운드 추가 했는데 시끄러우시면 오디오 클립만 빼시면 됌니당

+++ Assets/CSM/Enemy.cs
@@ -10,9 +10,9 @@ public class Enemy : MonoBehaviour, IDamageable
         currentHealth = maxHealth;
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
-        currentHealth -= amount;
+        currentHealth -= (int)amount;
 
         if (currentHealth <= 0)
         {
